immediate policy: optionally include past N weeks
Right now, the immediate policy uses cred from the most recent interval.  This generalizes
it such that users can create policies based on the past N weeks, without the discounting
of the Recent policy.

This includes an optional `numPeriodsLookback` parameter in the `ImmediatePolicy` type, which
defaults to 1 in the `immediateReceipts` method.  Keeping the parameter optional allows
backcompatibility, as users won't have to change their config to accommodate.

Note, this raises a question around naming, as Recent is actually a better name here, and we
could use something for the current Recent policies that highlights the discounting more.

__Test Plan__
Unit tests provided check that the `immediatePolicy` works for valid params, and also checks
for invalid `credIntervalCount` parameters.